requirements.txt
cohere
groq
python-dotenv
rich
opencv-python
googlesearch-python
requests
pillow
